Finalize structure of pizza, avocado smoothie, fried rice pages
I essentially finished the structure of the pizza, avocado smoothie, and
fried rice pages then connected them to the home page. I added pictures,
ingredients, and the steps to make them (off of various websites). My next
course of action is to probably add a picture to the home page and shrink
the sizes of the pictures in each recipe page (they are too big).