Resized images on all recipes and added an image on the home page
I finished resizing all the images in the recipe pages and added an image onto
the home page. I renamed the heading on the home page from Odin Recipes to
Lance's Recipes. While I got the recipe idea from the Odin Project, I felt
that as a programmer, I was the one who programmed the page from start to
finish. For now, I think this is the final version of the project until I
learn CSS and JavaScript!